+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Release NEXT
 - Emit resource state events for all resource models.
 - Fix events filtering if resource URL is specified as scope.
 - Fix ordering for /resources endpoint.
+- Implement pull operation for OpenStack tenant.
 
 Release 0.93.0
 --------------